Support register IPv6
