+++ spring-cloud-alibaba-examples/nacos-example/nacos-discovery-example/nacos-discovery-provider-example/src/main/resources/application.properties
@@ -3,6 +3,7 @@ spring.application.name=service-provider
 spring.cloud.nacos.discovery.server-addr=127.0.0.1:8848
 spring.cloud.nacos.discovery.enabled=true
 #spring.cloud.nacos.discovery.instance-enabled=true
+#register IPv6 instance
 #spring.cloud.nacos.discovery.ip-type=IPv6
 
 spring.cloud.nacos.username=nacos
