+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-nacos-discovery/src/test/java/com/alibaba/cloud/nacos/discovery/reactive/NacosReactiveDiscoveryClientConfigurationTests.java
@@ -17,6 +17,7 @@
 package com.alibaba.cloud.nacos.discovery.reactive;
 
 import com.alibaba.cloud.nacos.discovery.NacosDiscoveryAutoConfiguration;
+import com.alibaba.cloud.nacos.util.UtilIPv6AutoConfiguration;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
@@ -33,6 +34,7 @@ public class NacosReactiveDiscoveryClientConfigurationTests {
 
 	private ApplicationContextRunner contextRunner = new ApplicationContextRunner()
 			.withConfiguration(AutoConfigurations.of(UtilAutoConfiguration.class,
+					UtilIPv6AutoConfiguration.class,
 					NacosDiscoveryAutoConfiguration.class,
 					NacosReactiveDiscoveryClientConfiguration.class));
